Move standardization tests into ontology tests
