+++ indra/tests/test_ontology.py
@@ -1,6 +1,9 @@
+from indra.statements import Agent
 from indra.ontology.bio import bio_ontology
 from indra.ontology.world import world_ontology
 from indra.databases import go_client, hgnc_client
+from indra.ontology.standardize import \
+    standardize_agent_name, standardize_db_refs
 
 
 def test_isa_entity():
@@ -142,3 +145,134 @@ def test_hm_opposite_polarity():
     assert world_ontology.get_polarity('WM', concept1) == -1
     assert world_ontology.get_polarity('WM', concept2) == 1
     assert world_ontology.get_polarity('UN', 'something') is None
+
+
+def test_name_standardize_hgnc_up():
+    a1 = Agent('x', db_refs={'HGNC': '9387'})
+    standardize_agent_name(a1, True)
+    assert a1.name == 'PRKAG3'
+    a1 = Agent('x', db_refs={'UP': 'Q9UGI9'})
+    standardize_agent_name(a1, True)
+    assert a1.name == 'PRKAG3'
+    a1 = Agent('x', db_refs={'UP': 'Q8BGM7'})
+    standardize_agent_name(a1, True)
+    assert a1.name == 'Prkag3'
+
+
+def test_name_standardize_chebi():
+    a1 = Agent('x', db_refs={'CHEBI': 'CHEBI:15996'})
+    standardize_agent_name(a1, False)
+    assert a1.name == 'GTP'
+
+
+def test_name_standardize_go():
+    a1 = Agent('x', db_refs={'GO': 'GO:0006915'})
+    standardize_agent_name(a1, False)
+    assert a1.name == 'apoptotic process'
+
+
+def test_name_standardize_mesh():
+    a1 = Agent('x', db_refs={'MESH': 'D008545'})
+    standardize_agent_name(a1, False)
+    assert a1.name == 'Melanoma', a1.name
+
+
+def test_name_standardize_mesh_go():
+    a1 = Agent('x', db_refs={'MESH': 'D058750'})
+    standardize_agent_name(a1, True)
+    assert a1.db_refs['GO'] == 'GO:0001837'
+    assert a1.name == 'epithelial to mesenchymal transition', a1.name
+    a1 = Agent('x', db_refs={'GO': 'GO:0001837'})
+    standardize_agent_name(a1, True)
+    assert a1.db_refs['MESH'] == 'D058750'
+    assert a1.name == 'epithelial to mesenchymal transition', a1.name
+
+
+def test_name_standardize_mesh_other_db():
+    a1 = Agent('x', db_refs={'MESH': 'D001194'})
+    standardize_agent_name(a1, True)
+    assert a1.db_refs['CHEBI'] == 'CHEBI:46661'
+    assert a1.name == 'asbestos', a1.name
+
+    db_refs = {'MESH': 'D000067777'}
+    db_refs = standardize_db_refs(db_refs)
+    assert db_refs.get('HGNC') == '3313', db_refs
+    assert db_refs.get('UP') == 'Q12926', db_refs
+    a2 = Agent('x', db_refs=db_refs)
+    standardize_agent_name(a2)
+    assert a2.name == 'ELAVL2'
+
+
+def test_standardize_db_refs_efo_hp_doid():
+    refs = standardize_db_refs({'EFO': '0009502'})
+    assert refs.get('MESH') == 'D000007', refs
+    refs = standardize_db_refs({'MESH': 'D000007'})
+    assert refs.get('EFO') == '0009502', refs
+
+    refs = standardize_db_refs({'HP': 'HP:0031801'})
+    assert refs.get('MESH') == 'D064706', refs
+    refs = standardize_db_refs({'MESH': 'D064706'})
+    assert refs.get('HP') == 'HP:0031801', refs
+
+    # Currently there is no one-to-many mapping in the direction towards MeSH
+    # (there used to be) if there is again, we should test it here
+    #refs = standardize_db_refs({'DOID': 'DOID:0060695'})
+    #assert 'MESH' not in refs
+
+    # One-to-many mappings away from MESH
+    refs = standardize_db_refs({'MESH': 'D000071017'})
+    assert 'DOID' not in refs
+
+    refs = standardize_db_refs({'DOID': 'DOID:0060495'})
+    assert refs.get('MESH') == 'D000067208'
+
+    # This is an xrefs-based mapping that isn't in Gilda's resource file
+    refs = standardize_db_refs({'EFO': '0000694'})
+    assert refs.get('MESH') == 'D045169'
+
+
+def test_standardize_name_efo_hp_doid():
+    ag = Agent('x', db_refs={'HP': 'HP:0031801'})
+    standardize_agent_name(ag)
+    # Name based on MESH mapping
+    assert ag.name == 'Vocal Cord Dysfunction'
+
+    ag = Agent('x', db_refs={'HP': 'HP:0000002'})
+    standardize_agent_name(ag)
+    # Name based on HP itself
+    assert ag.name == 'Abnormality of body height'
+
+    ag = Agent('x', db_refs={'DOID': 'DOID:0014667'})
+    standardize_agent_name(ag)
+    # Name based on MESH mapping
+    assert ag.name == 'Metabolic Diseases'
+
+    ag = Agent('x', db_refs={'EFO': '1002050'})
+    standardize_agent_name(ag)
+    # Name based on MESH mapping
+    assert ag.name == 'Nephritis', (ag.name, ag.db_refs)
+
+    ag = Agent('x', db_refs={'EFO': '0000001'})
+    standardize_agent_name(ag)
+    # Name based on EFO itself
+    assert ag.name == 'experimental factor', (ag.name, ag.db_refs)
+
+
+def test_standardize_uppro():
+    ag = Agent('x', db_refs={'UP': 'P01019'})
+    standardize_agent_name(ag)
+    assert ag.name == 'AGT'
+    ag = Agent('x', db_refs={'UPPRO': 'PRO_0000032458'})
+    standardize_agent_name(ag)
+    assert ag.name == 'Angiotensin-2', ag.name
+    ag = Agent('x', db_refs={'UPPRO': 'PRO_0000032458', 'UP': 'P01019'})
+    standardize_agent_name(ag)
+    assert ag.name == 'Angiotensin-2', ag.name
+
+
+def test_uppro_fallback():
+    # This UP chain has no name currently so we can test that the fallback
+    # to naming by the UP ID is working
+    ag = Agent('x', db_refs={'UP': 'Q6IE75', 'UPPRO': 'PRO_0000383648'})
+    standardize_agent_name(ag)
+    assert ag.name == 'Bace2'
